# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateTaskRequest:
    properties:
      column_name:
        type: string
      description:
        type: string
      estimated_minutes:
        type: integer
      priority:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  models.Note:
    properties:
      content_json:
        type: string
      created_at:
        type: string
      id:
        type: string
      task_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Task:
    properties:
      column_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      estimated_minutes:
        type: integer
      id:
        type: string
      position:
        type: integer
      priority:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.UpdateNoteRequest:
    properties:
      content_json:
        type: string
    required:
    - content_json
    type: object
  models.UpdateTaskPositionRequest:
    properties:
      column_name:
        type: string
      position:
        type: integer
    required:
    - column_name
    type: object
  models.UpdateTaskRequest:
    properties:
      column_name:
        type: string
      description:
        type: string
      estimated_minutes:
        type: integer
      position:
        type: integer
      priority:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for task management with Kanban + Timeboxing + AI commands
  title: FLOWSTATE API
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: Returns all tasks for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Task'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: 'Creates a task in the specified Kanban column. Default column
        is "backlog". Priority options: low, medium, high.'
      parameters:
      - description: Task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Deletes a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      description: Returns a specific task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Updates an existing task with new data
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task data to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
  /tasks/{id}/complete:
    post:
      description: Marks a task as completed by moving it to the "done" column. Triggers
        any associated AI workflow notifications.
      parameters:
      - description: ID of the task
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete task
      tags:
      - tasks
  /tasks/{id}/note:
    get:
      description: Returns the note associated with a task (singleton per task). Returns
        empty note if none exists.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get task note
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Creates or replaces the note for a task (idempotent PUT)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Note content JSON
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Save task note
      tags:
      - notes
  /tasks/{id}/position:
    patch:
      consumes:
      - application/json
      description: Updates task position and column (drag & drop)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: New position data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskPositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Move task
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"